Breaking driverOS into multiple OS files - one for each user type
