+++ driverOS.cpp
@@ -1,5 +1,6 @@
 /*
-Implementation file for driver.cpp
+Nathan Obert M03134502
+Implementation file for driver.cpp, further broken down with each User Type having its own OS .cpp file
 This is a way of seperating functions that are used inside the driver to manipulate
 the user, account, and object data.
 */
@@ -10,24 +11,9 @@ bool isValidOption(string input, int upperBound);
 int getUserOption(int upperBound);
 void userLoginReset();
 
-void clientLogin(string userID);
-void clientMakeAccountChanges(Client &user, int option);
-void clientHelpAccessAccount(Client &user, int option, vector<string> acctList);
-
-void officialLogin(string userID);
-void officialOpenAccounts(Official &officialUser);
-void officialSearch(Official &officialUser);
-
-void adminLogin(string userID);
-void adminModifyOfficial(Admin &admin);
-void adminModifyAdmin(Admin &admin);
-void adminModifyAccountTypes(Admin &admin);
-void adminViewTableStats(Admin &admin);
-void adminRetrieveUserID(Admin &admin);
-
 
 /************************************************
-Initial helper functions
+Initial helper functions for main
 ************************************************/
 
 
@@ -263,943 +249,3 @@ void userLoginReset()
     }
 }
 
-
-/************************************************
-Start of Client Login
-************************************************/
-
-
-//Needs Work still!!
-void clientLogin(string userID)
-{
-    Client user;
-    user.buildUser("UserData/" + userID + ".txt");
-    cout << "Welcome, " << user.getName() << endl;
-    cout << "Last Activity: " << user.getRecentActivity() << endl;
-    cout << "Last Login: " << user.getRecentLogin() << endl << endl;
-    user.setRecentLogin(DateTools().getCurrentTime()); //since we just logged in, now need to update time
-
-    string clientInterface = "[1] Access Accounts\n[2] View Personal Information\n[3] Change Information\n[4] Open New Account\n[5] Log Out";
-    bool wantsToExit = false;
-    
-    while (!wantsToExit)
-    {
-        cout << clientInterface << endl << "Option: ";
-        int initialOption = getUserOption(5);
-        cout << endl;
-
-        switch (initialOption)
-        {
-            case 1: //Accounts
-            {
-                bool doAccountsExist = DataHandler::clientDisplayAccounts(userID);
-
-                if (!doAccountsExist) //won't display anything
-                {
-                    cout << "There is no account data to access" << endl << endl;
-                }
-                else //accounts have been displayed by DataHandler already
-                {
-                    string accountInterface = "[1] Deposit Into Account\n[2] Withdraw From Account\n[3] Deposit Into External Account\n[4] View an Accounts History\n[5] Go Back";
-                    cout << accountInterface << endl << "Option: ";
-                    int accountInterfaceOption = getUserOption(5);
-                    cout << endl;
-                    
-                    if (accountInterfaceOption != 5)
-                    {
-                        vector<string> acctList = DataHandler::clientGetAccountList(userID);
-                        clientHelpAccessAccount(user, accountInterfaceOption, acctList);
-                    }
-                }
-                break;
-            }
-            case 2: //Display user info, i.e. name, address, phone num
-            {
-                user.displayInfo();
-                user.setRecentActivity("Displayed Personal Information");
-                break;
-            }
-            case 3: //change info
-            {
-                cout << "Which piece of information would you like to change?" << endl;
-                cout << "[1] First Name\n[2] Last Name\n[3] Address\n[4] Phone Number\n[5] Password\n[6] Go Back" << endl;
-                cout << "Option: ";
-                int option = getUserOption(6);
-                if (option != 6)
-                {
-                    cout << endl;
-                    clientMakeAccountChanges(user, option);
-                }
-                cout << endl;
-                break;
-            }
-            case 4: //open new acct
-            {
-                string acctType = "a cool account type";
-                //Make sure to display all types of accts currently offered
-                //Upon user choosing, send request off to an official for confirmation!
-                //have all officials be able to access this list of requests!
-                //Each request needs: UserID requesting and acct type
-                //Upon accpeting request or denying request for new acct, update client recent activity message to display the choice
-
-                DataHandler::clientRequestNewAccount(userID, acctType);
-                user.setRecentActivity("Requested New Account: " + acctType);
-                cout << "Please be aware that a Bear Bank Official may take several business days to review this request." << endl << endl;
-                break;
-            }
-            case 5: //Log out
-            {
-                wantsToExit = true;
-                user.saveUser();
-            }
-        }
-    }
-}
-
-//Needs Finishing!
-void clientMakeAccountChanges(Client &user, int option)
-{
-    switch (option)
-    {
-        case 1: //change first name
-        {
-            string name = "", oldName = user.getName(), lastName = user.getName(), userID = user.getID();
-            oldName = oldName.substr(0, oldName.find(" "));
-            lastName = lastName.substr(lastName.find(" ") + 1, string::npos);
-
-            //changing name in user obj
-            cout << "Enter your new first name: ";
-            getline(cin, name);
-            name = name + " " + lastName;
-            user.setName(name);
-            user.setRecentActivity("Changed First Name");
-
-            //need to reflect changes on all accts and tables as well
-            DataHandler::changeClientFirstName(userID, oldName, name); //needs finishing
-
-            break;
-        }
-        case 2: //change last name
-        {
-            string firstName = user.getName(), name = "", oldName = user.getName(), userID = user.getID();
-            firstName = firstName.substr(0, firstName.find(" "));
-            oldName = oldName.substr(oldName.find(" ") + 1, string::npos);
-
-            //changing name in user obj
-            cout << "Enter your new last name: ";
-            getline(cin, name);
-            name = firstName + " " + name;
-            user.setName(name);
-            user.setRecentActivity("Changed Last Name");
-
-            //changing name inside the tables
-            DataHandler::changeClientLastName(userID, oldName, name); //needs finishing
-
-            break;
-        }
-        case 3: //change address
-        {
-            string newAddress = "", oldAddress = user.getAddress();
-
-            //chaning address in user obj
-            cout << "Enter your new address: ";
-            getline(cin, newAddress);
-            user.setAddress(newAddress);
-            user.setRecentActivity("Changed Address");
-
-            //changing address in table
-            DataHandler::changeClientAddress(user.getID(), oldAddress, newAddress); //needs finishing
-
-            break;
-        }
-        case 4: //change phone number
-        {
-            string newNum = "", oldNum = user.getPhoneNum();
-
-            //chaning address in user obj
-            cout << "Enter your new phone number: ";
-            getline(cin, newNum);
-            user.setAddress(newNum);
-            user.setRecentActivity("Changed Phone Number");
-
-            //changing address in table
-            DataHandler::changeClientPhoneNum(user.getID(), oldNum, newNum); //needs finishing
-            break;
-        }
-        case 5: //change password
-        {
-            string newPassword = "";
-            cout << "Enter your new password: ";
-            getline(cin, newPassword);
-
-            Admin admin; //creating the Automated Admin obj to change the password
-            admin.buildUser("UserData/admin.txt");
-            admin.setRecentLogin(DateTools().getCurrentTime()); //setting most recent login time as now
-            admin.resetPassword(user.getID(), newPassword);
-            admin.setRecentActivity("Assisted Client '" + user.getID() + "' Change Password in Settings");
-
-            user.setRecentActivity("Password was Reset in Settings");
-            user.setPassword(EncryptionBox::hash(newPassword)); //reflecting change inside user obj
-            user.saveUser();
-            cout << "Your password has been reset." << endl;
-            break;
-        }
-    }
-    user.saveUser();
-}
-
-//Needs Finishing!
-void clientHelpAccessAccount(Client &user, int option, vector<string> acctList)
-{
-    switch (option)
-    {
-        case 1: //deposit
-        {
-            break;
-        }
-        case 2: //withdraw
-        {
-            break;
-        }
-        case 3: //deposit into another acct in Bear Bank
-        {
-            break;
-        }
-        case 4: //view history
-        {
-            user.getAccountHistory(acctList); //Needs to be finished!!!
-            break;
-        }
-    }
-}
-
-
-/************************************************
-Start of Official Login
-************************************************/
-
-
-void officialLogin(string userID)
-{
-    Official user;
-    user.buildUser("UserData/" + userID + ".txt");
-
-    if (user.getState() != "active")
-    {
-        cout << "This account has been locked." << endl << endl;
-    }
-    else
-    {
-        cout << "Welcome, " << user.getName() << endl;
-        cout << "Last Activity: " << user.getRecentActivity() << endl;
-        cout << "Last Login: " << user.getRecentLogin() << endl << endl;
-        user.setRecentLogin(DateTools().getCurrentTime());
-
-        string officialInterface = "[1] Open Account\n[2] Close Account\n[3] Alter Existing Account\n[4] Deposit into Account\n[5] Withdraw from Account\n[6] Search for Bank Accounts\n[7] Log Out";
-        bool wantsToExit = false;
-        
-        while (!wantsToExit)
-        {
-            cout << officialInterface << endl << "Option: ";
-            int initialOption = getUserOption(7);
-            cout << endl;
-
-            switch (initialOption)
-            {
-                case 1: //Open acct - look at queue or just open new acct for specified user
-                {
-                    officialOpenAccounts(user); //abstracted due to needed while loop
-                    break;
-                }
-                case 2: //Close acct
-                {
-                    break;
-                }
-                case 3: //Alter existing acct - changing things like interest rate, term length, maturity date, etc
-                {
-                    //build acct
-                    //check if open
-                    //make changes, if applicable to acct type
-                    //update DataHandler.allTables.accountTable to have new info
-                    //save acct
-                    break;
-                }
-                case 4: //Deposit into acct - needs user confirmation
-                {
-                    break;
-                }
-                case 5: //Withdraw from acct - needs user confirmation
-                {
-                    break;
-                }
-                case 6: //Search for acct by userID, first name, last name, phone num, address, acct num
-                {
-                    officialSearch(user);
-                    break;
-                }
-                case 7: //Exit
-                {
-                    wantsToExit = true;
-                    user.saveUser();
-                    break;
-                }
-            }
-        }
-    }
-}
-
-void officialOpenAccounts(Official &officialUser)
-{
-    string openAcctInterface = "[1] Open a new Account\n[2] View User Requested Accounts\n[3] Go Back";
-    string requestedAcctInterface = "[1] Approve Request\n[2] Deny Request\n[3] Go Back";
-    bool wantsToExit = false;
-    
-    while (!wantsToExit)
-    {
-        cout << openAcctInterface << endl << "Option: ";
-        int mainOption = getUserOption(3);
-        cout << endl;
-        
-        switch (mainOption)
-        {
-            case 1: //Open new acct
-            {
-                //get userID where acct will be added
-                //display acct types
-                //get selection
-                //add acct to userTable and accountTable 
-
-                //make all of ^ into official funct!!!
-                break;
-            }
-            case 2: //View acct request queue
-            {
-                AccountQueue queue;
-                queue.buildQueue();
-                string request = "", userID = "", acctType = "";
-                bool doneReviewing = false;
-
-                if (queue.isEmpty())
-                {
-                    cout << "There are no account openning requests to view!" << endl << endl;
-                    queue.saveQueue();
-                    break;
-                }
-                else
-                {
-                    while (!doneReviewing)
-                    {
-                        queue.peekFirst(request); //request takes on first item in queues value, queue retains order
-                        userID = request.substr(0, request.find(" ")); //request is formatted: "<userID>  requests a new: <acctType>"
-                        acctType = request.substr(request.find(":") + 2, string::npos); //finds the : inside "a new: " and gets whatever is after
-                        Client clientUser;
-                        clientUser.buildUser("UserData/" + userID + ".txt");
-
-                        cout << request << endl;
-
-                        cout << requestedAcctInterface << endl << "Option: ";
-                        int reviewOption = getUserOption(3);
-
-                        switch (reviewOption)
-                        {
-                            case 1: //Approve
-                            {  
-                                string acctNum = "", dummyStr = "";
-                                //create a new acct obj of specified type
-                                //Account newAccount;
-                                //save new acct obj to create a record in files
-                                //update acctNum!
-
-                                //DataHandler::addClientAccountToRecords(user, newAccount); //updates all tables with new info!
-                                clientUser.setRecentActivity("Request for: " + acctType + " was approved!");
-                                clientUser.saveUser();
-                                officialUser.setRecentActivity("Approved Request from: " + userID + " for: " + acctType);
-                                officialUser.saveUser();
-                                queue.dequeue(dummyStr); //dequeue needs a string& passed, so this just allows a value to be removed
-                                cout << "Approved Request" << endl;
-
-                                cout << endl;
-                                break;
-                            }
-                            case 2: //Deny
-                            {
-                                string dummyStr = "";
-                                clientUser.setRecentActivity("Request for: " + acctType + " was denied.");
-                                clientUser.saveUser();
-                                officialUser.setRecentActivity("Denied Request from: " + userID + " for: " + acctType);
-                                officialUser.saveUser();
-                                queue.dequeue(dummyStr);
-                                cout << "Denied Request" << endl;
-
-                                cout << endl;
-                                break;
-                            }
-                            case 3: //Go Back
-                            {
-                                doneReviewing = true;
-                                //save queue back to file
-                                queue.saveQueue();
-                                cout << endl;
-                                break;
-                            }
-                        }
-                        if (queue.isEmpty()) //break condiditon while reviewing
-                        {
-                            cout << "There are no more Requests to View" << endl << endl;
-                            doneReviewing = true;
-                            queue.saveQueue();
-                            break;
-                        }
-                    }
-                }
-
-                break;
-            }
-            case 3: //Go back
-            {
-                wantsToExit = true;
-                break;
-            }
-        }
-    }
-}
-
-void officialSearch(Official &officialUser)
-{
-    string searchInterface = "[1] Search by User ID\n[2] Search by First Name\n[3] Search by Last Name\n[4] Search by Account Number\n[5] Search by Phone Number\n[6] Search by Address\n[7] View a Closed Account\n[8] Go Back";
-    bool wantsToExit = false;
-
-    while (!wantsToExit)
-    {
-        cout << searchInterface << endl << "Option: ";
-        int searchOption = getUserOption(8);
-
-        switch (searchOption)
-        {
-            case 1: //user id
-            {
-                string userID = "";
-                cout << "Enter a User ID: ";
-                getline(cin, userID);
-                cout << endl;
-                bool isValid = DataHandler::isValidUserID(userID);
-
-                if (isValid)
-                {
-                    officialUser.searchByUserID(userID);
-                    officialUser.setRecentActivity("Searched for Bank Accounts by a User ID");
-                    officialUser.saveUser();
-                }
-                else
-                {
-                    cout << "The Entered User ID is Invalid" << endl;
-                }
-                
-                break;
-            }
-            case 2: //first name
-            {
-                string firstName = "";
-                cout << "Enter a first name: ";
-                getline(cin, firstName);
-                cout << endl;
-                officialUser.searchByFirstName(firstName);
-                officialUser.setRecentActivity("Searched for Bank Accounts by a First Name");
-                officialUser.saveUser();
-                break;
-            }
-            case 3: //last name
-            {
-                string lastName = "";
-                cout << "Enter a last name: ";
-                getline(cin, lastName);
-                cout << endl;
-                officialUser.searchByLastName(lastName);
-                officialUser.setRecentActivity("Searched for Bank Accounts by a Last Name");
-                officialUser.saveUser();
-                break;
-            }
-            case 4: //acct num
-            {
-                string acctNum = "";
-                cout << "Enter an Account Number: ";
-                getline(cin, acctNum);
-                cout << endl;
-                officialUser.searchByAccountNum(acctNum);
-                officialUser.setRecentActivity("Searched for a Bank Account by an Account Number");
-                officialUser.saveUser();
-                break;
-            }
-            case 5: //phone num
-            {
-                string phoneNum = "";
-                cout << "Enter a phone number: ";
-                getline(cin, phoneNum);
-                cout << endl;
-                officialUser.searchByPhoneNum(phoneNum);
-                officialUser.setRecentActivity("Searched for Bank Accounts by a Phone Number");
-                officialUser.saveUser();
-                break;
-            }
-            case 6: //address
-            {
-                string address = "";
-                cout << "Enter an address: ";
-                getline(cin, address);
-                cout << endl;
-                officialUser.searchByAddress(address);
-                officialUser.setRecentActivity("Searched for Bank Accounts by an Address");
-                officialUser.saveUser();
-                break;
-            }
-            case 7: //closed acct display -- must implement!!
-            {
-                //get acct num
-                //display short into
-                //display who closed by
-                //display when closed
-                break; 
-            }
-            case 8: //go back
-            {
-                wantsToExit = true;
-                cout << endl;
-                break;
-            }
-        }
-    }
-}
-
-
-
-/************************************************
-Start of Admin Login
-************************************************/
-
-
-void adminLogin(string userID)
-{
-    Admin admin;
-    admin.buildUser("UserData/" + userID + ".txt");
-    cout << "Welcome, " << admin.getName() << endl;
-    cout << "Last Activity: " << admin.getRecentActivity() << endl;
-    cout << "Last Login: " << admin.getRecentLogin() << endl << endl;
-    admin.setRecentLogin(DateTools().getCurrentTime());
-
-    string adminInterface = "[1] Modify Official Account\n[2] Modify an Admin Account\n[3] Modify Account Types\n[4] Retrieve a User ID\n[5] Change a Password\n[6] View Table Statistics\n[7] Log Out";
-    bool wantsToExit = false;
-
-    while(!wantsToExit)
-    {
-        cout << adminInterface << endl << "Option: ";
-        int initialOption = getUserOption(7);
-        cout << endl;
-
-        switch (initialOption)
-        {
-            case 1: //Modify Official Account
-            {
-                adminModifyOfficial(admin);
-                break;
-            }
-            case 2: //Modify Admin
-            {
-                adminModifyAdmin(admin);
-                break;
-            }
-            case 3: //Modify Account Types
-            {
-                adminModifyAccountTypes(admin);
-                break;
-            }
-            case 4: //Retrieve User ID
-            {
-                adminRetrieveUserID(admin);
-                break;
-            }
-            case 5: //Change pw
-            {
-                string userName = "", newPassword = "", userType = "";
-                cout << "Enter the User ID: ";
-                getline(cin, userName);
-
-                if (DataHandler::isValidUserID(userName))
-                {
-                    cout << "Enter the New Password: ";
-                    getline(cin, newPassword);
-                    admin.resetPassword(userName, newPassword);
-                    admin.setRecentActivity("Reset Password for: " + userName);
-                    admin.saveUser();
-
-                    //now to reflect update in other user acct
-                    userType = DataHandler::getLoginInfo(userName)[1];
-
-                    if (userType == "client")
-                    {
-                        Client user;
-                        user.buildUser("UserData/" + userName + ".txt");
-                        user.setRecentActivity("Password reset by: " + userID);
-                        user.saveUser();
-                    }
-                    else if (userType == "official")
-                    {
-                        Official user;
-                        user.buildUser("UserData/" + userName + ".txt");
-                        user.setRecentActivity("Password reset by: " + userID);
-                        user.saveUser();
-                    }
-                    else
-                    {
-                        Admin user;
-                        user.buildUser("UserData/" + userName + ".txt");
-                        user.setRecentActivity("Password reset by: " + userID);
-                        user.saveUser();
-                    }
-                    cout << endl << "The Password has been Changed!" << endl;
-                }
-                else
-                {
-                    cout << "Invalid User ID Entered" << endl << endl;
-                }
-                
-                break;
-            }
-            case 6: //view statistics
-            {
-                adminViewTableStats(admin);
-                break;
-            }
-            case 7: //Log Out
-            {
-                wantsToExit = true;
-                admin.saveUser();
-                break;
-            }
-        }
-    }
-}
-
-void adminModifyOfficial(Admin &admin)
-{
-    string modifyOfficialInterface = "[1] Create a New Official\n[2] Edit an Existing Official\n[3] Go Back";
-    bool wantsToExit = false;
-
-    while (!wantsToExit)
-    {
-        cout << modifyOfficialInterface << endl << "Option: ";
-        int modifyOfficialOption = getUserOption(3);
-        cout << endl;
-
-        switch (modifyOfficialOption)
-        {
-            case 1: //Create New Official
-            {
-                string fullName = "", userID = "", password = "";
-                bool isUserIDAvailable = false;
-                cout << "Enter the Official's first and last name: ";
-                getline(cin, fullName);
-
-                while (!isUserIDAvailable)
-                {
-                    cout << "Enter the Official's User ID: ";
-                    getline(cin, userID);
-                    isUserIDAvailable = DataHandler::isAvaliableUserID(userID);
-                    if (!isUserIDAvailable)
-                    {
-                        cout << "Entered User ID not Available" << endl;
-                    }
-                }
-                
-                cout << "Enter the Official's password: ";
-                getline(cin, password);
-
-                admin.createOfficial(fullName, userID, password);
-                admin.setRecentActivity("Created new official: " + userID);
-                admin.saveUser();
-                cout << "A New Official Account for: " << fullName << " has been created!" << endl << endl;                
-                break;
-            }
-            case 2: //Edit Existing official
-            {
-                string officialUserID = "";
-                cout << "Enter the User ID of the official: ";
-                getline(cin, officialUserID);
-                cout << endl;
-
-                if (officialUserID == "official")
-                {
-                    cout << "Cannot edit the Automated Official" << endl;
-                }
-                else
-                {
-                    vector<string> userInfo = DataHandler::clientGetAccountList(officialUserID);
-
-                    if (userInfo.size() == 0) //no results found from searching, i.e. returned vect was {}
-                    {
-                        cout << "User ID not found" << endl << endl;
-                    }
-                    else if (userInfo[1] != "official") 
-                    {
-                        cout << officialUserID << " is not a Bear Bank Official" << endl << endl;
-                    }
-                    else //found an official!
-                    {
-                        cout << "Official: " << officialUserID << " was found!" << endl << endl;
-                        Official officialUser;
-                        officialUser.buildUser("UserData/" + officialUserID + ".txt");
-
-                        string editOfficialInterface = "[1] Set Active\n[2] Set Inactive\n[3] Delete Official\n[4] Go Back";
-                        cout << editOfficialInterface << endl << "Option: ";
-                        int editOfficialOption = getUserOption(4);
-                        cout << endl;
-
-                        switch (editOfficialOption)
-                        {
-                            case 1: //Set active
-                            {
-                                admin.setOfficialActive(officialUser);
-                                admin.setRecentActivity("Set Official Active: " + officialUserID);
-                                admin.saveUser();
-                                cout << officialUserID << " has been set to: Active" << endl;
-                                cout << endl;
-                                break;
-                            }
-                            case 2: //Set inactive
-                            {
-                                admin.setOfficialInactive(officialUser);
-                                admin.setRecentActivity("Locked out Official: " + officialUserID);
-                                admin.saveUser();
-                                cout << officialUserID << " has been set to: Inactive" << endl;
-                                cout << endl;
-                                break;
-                            }
-                            case 3: //Delete official
-                            {
-                                admin.deleteOfficial(officialUser);
-                                admin.setRecentActivity("Deleted Official: " + officialUserID);
-                                admin.saveUser();
-                                cout << officialUserID << " has been deleted" << endl;
-                                cout << endl;
-                                break;
-                            }
-                            case 4: //Go back
-                            {
-                                break;
-                            }
-                        }
-                    }
-                }
-
-                break;
-            }
-            case 3: //Go back
-            {
-                wantsToExit = true;
-                break;
-            }
-        }
-    }
-}
-
-void adminModifyAdmin(Admin &admin)
-{
-    string modifyAdmin = "[1] Create a New Admin\n[2] Delete an Existing Admin\n[3] Go Back";
-    bool wantsToExit = false;
-
-    while (!wantsToExit)
-    {
-        cout << modifyAdmin << endl << "Option: ";
-        int modifyAdminOption = getUserOption(3);
-        cout << endl;
-
-        switch (modifyAdminOption)
-        {
-            case 1: //new admin
-            {
-                string fullName = "", userID = "", password = "";
-                bool isUserIDAvailable = false;
-                cout << "Enter the Admin's first and last name: ";
-                getline(cin, fullName);
-
-                while (!isUserIDAvailable)
-                {
-                    cout << "Enter the Admin's User ID: ";
-                    getline(cin, userID);
-                    isUserIDAvailable = DataHandler::isAvaliableUserID(userID);
-                    if (!isUserIDAvailable)
-                    {
-                        cout << "Entered User ID not Available" << endl;
-                    }
-                }
-                
-                cout << "Enter the Admin's password: ";
-                getline(cin, password);
-                admin.createAdmin(fullName, userID, password);
-                admin.setRecentActivity("Created new admin: " + userID);
-                admin.saveUser();
-                cout << endl << "A New Admin Account for: " << fullName << " has been created!" << endl << endl;                
-                break;
-            }
-            case 2: //delete existing admin
-            {
-                string userID = "";
-                cout << "Enter the User ID of the Admin you want to Delete: ";
-                getline(cin, userID);
-                vector<string> userInfo = DataHandler::getLoginInfo(userID);
-
-                if (userID == "admin")
-                {
-                    cout << endl << "Cannot delete the Automated Admin" << endl << endl;
-                }
-                else if (userInfo.size() == 0)
-                {
-                    cout << endl << "Invalid User ID" << endl << endl;
-                }
-                else if (userInfo[1] != "admin")
-                {
-                    cout << endl << "User ID does not Belong to an Admin" << endl << endl;
-                }
-                else
-                {
-                    admin.deleteAdmin(userID);
-                    admin.setRecentActivity("Deleted admin: " + userID);
-                    admin.saveUser();
-                    cout << endl << "The Admin Account: " + userID + " has been deleted" << endl << endl;
-                }
-                break;
-            }
-            case 3:
-            {
-                wantsToExit = true;
-                break;
-            }
-        }
-    }
-}
-
-void adminModifyAccountTypes(Admin &admin)
-{
-    string modifyAcctTypeInterface = "[1] View All Account Types\n[2] Create New Account Type\n[3] Delete Existing Account Type\n[4] Go Back";
-    bool wantsToExit = false;
-
-    while (!wantsToExit)
-    {
-        cout << modifyAcctTypeInterface << endl << "Option: ";
-        int modifyAcctTypeOption = getUserOption(4);
-        cout << endl;
-        
-        switch (modifyAcctTypeOption)
-        {
-            case 1: //View all
-            {
-                break;
-            }
-            case 2: //Create new
-            {
-                break;
-            }
-            case 3: //Delete existing
-            {
-                break;
-            }
-            case 4:
-            {
-                wantsToExit = true;
-                break;
-            }
-        }
-    }
-}
-
-void adminViewTableStats(Admin &admin)
-{
-    string tableStatsInterface = "[1] View User Table Statistics\n[2] View First Name Table Statistics\n[3] View Last Name Table Statistics\n[4] View Phone Number Table Statistics\n[5] View Address Table Statistics\n[6] View Account Table Statistics\n[7] Go Back";
-    bool wantsToExit = false;
-
-    while (!wantsToExit)
-    {
-        cout << tableStatsInterface << endl << "Option: ";
-        int tableStatsOption = getUserOption(7);
-        
-        switch (tableStatsOption)
-        {
-            case 1:
-            {
-                DataHandler::allTables.userTable.getStatistics();
-                break;
-            }
-            case 2: //First name
-            {
-                DataHandler::allTables.firstNameTable.getStatistics();
-                break;
-            }
-            case 3: //Last name
-            {
-                DataHandler::allTables.lastNameTable.getStatistics();
-                break;
-            }
-            case 4: //Phone num
-            {
-                DataHandler::allTables.phoneNumTable.getStatistics();
-                break;
-            }
-            case 5: //Address
-            {
-                DataHandler::allTables.addressTable.getStatistics();
-                break;
-            }
-            case 6: //Accounts
-            {
-                DataHandler::allTables.accountTable.getStatistics();
-                break;
-            }
-            case 7: //Go back
-            {
-                wantsToExit = true;
-                break;
-            }
-        }
-        cout << endl;
-    }
-}
-
-void adminRetrieveUserID(Admin &admin)
-{
-    string retrieveIDInterface = "[1] Display All Official User IDs\n[2] Display All Admin User IDs\n[3] Search for Client User ID by an Account Number\n[4] Go Back";
-    bool wantsToExit = false;
-    while (!wantsToExit)
-    {
-        cout << retrieveIDInterface << endl << "Option: ";
-        int retrieveIDOption = getUserOption(4);
-        cout << endl;
-
-        switch (retrieveIDOption)
-        {
-            case 1: //display all Officials
-            {
-                admin.displayAllOfficials();
-                admin.setRecentActivity("Displayed All Official User IDs");
-                admin.saveUser();
-                cout << endl;
-                break;
-            }
-            case 2: //display all Admins
-            {
-                admin.displayAllAdmins();
-                admin.setRecentActivity("Displayed All Admin User IDs");
-                admin.saveUser();
-                cout << endl;
-                break;
-            }
-            case 3: //Search For Client by acct num
-            {
-                break;
-            }
-            case 4:
-            {
-                wantsToExit = true;
-                break;
-            }
-        }
-    }
-}
